refactor(vector): search only via pgvector operators, move in-memory store to tests
Address review feedback that similarity search must use DB operators, not
hand-rolled JS. The shipped library now performs vector search exclusively
through pgvector.

- remove MemoryVectorStore + cosineSimilarity + matchesObjectFilter from
  src/vector.ts; the shipped store (PgVectorStoreImpl) searches only via the
  pgvector <=> operator plus JSONB @> filtering, all in Postgres
- move the in-memory cosine store and its provider to __tests__/fixtures as a
  test-only double so unit tests need no live pgvector
- drop the instanceof MemoryVectorStore branch in the dataset/template
  retrievers (the object filter drives the same SQL for both)
- stop exporting the in-memory store from the package (dev/test only; the host
  uses the Postgres store)
