+++ src/__tests__/fixtures/inmemory-vector-store.provider.ts
@@ -2,8 +2,9 @@ import {inject, Provider, service, ValueOrPromise} from '@loopback/core';
 import {AiIntegrationBindings} from '../../keys';
 import {EmbeddingService} from '../../services/embedding.service';
 import {EmbeddingProvider} from '../../types';
-import {MemoryVectorStore} from '../../vector';
+import {MemoryVectorStore} from './memory-vector-store';
 
+/** Test-only provider binding the in-memory vector store (see test-app). */
 export class InMemoryVectorStore implements Provider<MemoryVectorStore> {
   constructor(
     @service(EmbeddingService)
