+++ src/components/db-query/providers/templates.retriever.ts
@@ -1,10 +1,5 @@
 import {inject, Provider, ValueOrPromise} from '@loopback/core';
-import {
-  BaseRetriever,
-  Document,
-  MemoryVectorStore,
-  VectorStore,
-} from '../../../vector';
+import {BaseRetriever, VectorStore} from '../../../vector';
 import {AiIntegrationBindings} from '../../../keys';
 import {DbQueryStoredTypes} from '../types';
 import {AuthenticationBindings} from 'loopback4-authentication';
@@ -18,15 +13,6 @@ export class TemplateRetriever implements Provider<BaseRetriever> {
     private readonly user: IAuthUserWithPermissions,
   ) {}
   value(): ValueOrPromise<BaseRetriever> {
-    if (this.vectorStore instanceof MemoryVectorStore) {
-      return this.vectorStore.asRetriever({
-        k: 5,
-        filter: (doc: Document) =>
-          doc.metadata.type === DbQueryStoredTypes.Template &&
-          doc.metadata.tenantId === this.user.tenantId,
-        searchType: 'similarity',
-      });
-    }
     return this.vectorStore.asRetriever({
       k: 5,
       filter: {
