+++ src/providers/index.ts
@@ -1,3 +1,2 @@
 export * from './cache';
 export * from './tools.provider';
-export * from './vector-stores';
